Some basic seeds. Users, Activities. Go nuts
There are probably more seeds that could be added. This is enough that
when you login you can see events on basically every day, and there are
students for each cohort (one of which is current).

You still need to sign up and set yourself to be a student/teacher via
console... that could be an improvement.