Enhance assistance request button for students; addresses issues #120, #156, and #158.
Update markup in navbar. The navbar markup is sufficiently complex/erroneous that a future refactor should be done to clean it up. IDs were added to better attach behaviours to assistance request elements and to differentiate them from content on the assistance_requests page. Styles were also updated to be slightly more correct, but will likely be made redundant by a UI review.

Assistance request button behaviours were moved from the assistance_requests page's coffeescript file to a new navigation coffeescript file. In addition to the existing click behaviours, a polling mechanism was added to live-update the state of the assistance request button. A route was added to convey the status of the latest pertinent assistance request for students. It returns the state of the latest request - either no active request, active unclaimed request along with position in queue (new method on AssistanceRequest), or active request that was claimed along with the name of the teacher who claimed it.

To address the issue of multiple assistance requests being active for a student, a before_create callback was added that only allows the new assistance request to be created if there is not another active request. This way, even if a student manages to mash the button, they will only have one assistance request from the teacher side.