Remove User methods that are only used once and don't add value.
User#assistance_currently_requested_or_in_progress? was being used in the navigation when the Request Assitance button only had two states. Now that it has three states, the method isn't useful anymore.

User#latest_open_assistance_request was an idea that was being toyed with for the workflow of the new Assistance Request button enhancements, but doesn't add value.