Allow students to cancel in progress assistance requests.
Previously students could only cancel open assistance requests. This left the door open for a teacher to forget to complete an assistance and then the student would not be able to request further assistance.

The Assistance.currently_active scope had to be modified to exclude Assistances whose AssistanceRequests have been canceled.