Simplify and improve assistance request button UI.
Now that in progress assistances are cancelable by students, there are two UI states for the button: "able to create a new request" and "able to cancel current request" which are mutually exclusive.

New method User#position_in_queue returns either the current user's position in the assistance request queue or nil.