Enable teachers to select cohort locations on request queue.
A cookie is set when a teacher updates their preference for which cohort locations they are assisting. This is done using JavaScript on the assistance requests index page.

This exposes an issue with the design of the request queue where students see their position in the "global" queue. That is, their AssistanceRequst#position_in_queue takes into account cohorts in all locations.