Add AWS S3 file access for recordings.
Add aws-sdk-core gem for generating links.

Add new environment variables to .env.example file to indicate required ENV for prod.

Add recordings_folder and recordings_bucket to Program to allow for the videos associated with different programs to be stored on the same AWS account. recordings_folder is optional, and the object key of the file is just the file name if no folder is specified. Bad things will happen if there is no recordings_bucket or if there is no cohort associated with the recording or if that cohort doesn't have a program!

Add public s3_url method to Recording and use the method in pages that show links to watch or download recordings.