Fix main layout columns. Bulletproof activity list.
Rather than using the grid system, create fixed size columns that adapt based on the container size.

Some class names were changed in order to discover style conflicts. These can be changed to more correct class names in the future.

A fair bit of gardening was done for this commit. Sorry for the noise.